requirements.txt
requests>=2.31.0
yfinance>=0.2.18
PyPDF2>=3.0.1
python-dotenv>=1.0.0